Defender and Crusher tail smash has a 1% chance of doing a BONK! sound instead. (#3750)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->
Defender and Crusher tail smash has a 1% chance of doing a BONK! sound
instead.

# Explain why it's good for the game
its one of those fun occasional things it will be cool for gameplay 

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Defender Tail Slam has a 1% chance to do a BONK sound instead.
/:cl: