New Varadero CO Survivor (#3799)
# About the pull request

Adds a USASF Commander as a CO survivor to New Varadero. The idea is the
commander was visiting the outpost when the xeno infestation started.
Lore and flavor is up for grabs if someone has a better way to make the
role naval.

# Explain why it's good for the game

It's neat.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
add: Added a visiting USASF Commander (CO survivor) to New Varadero
/:cl: