Create humans tweaks (#3812)
# About the pull request

This PR adds extra length to the create humans menu so you don't have to
scroll every time.

This PR sets the default "range to spawn" variable to 0 as we don't
really want variance unless it is an active decision to do so.

# Explain why it's good for the game

Easier usage of admin tools good.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
qol: Create humans tab length increase
qol: Create humans tab now defaults to 0 range to spawn
/:cl: