Fixes extraneous messages in regards to wall nests (#3807)
# About the pull request

Mouse drop was not checking for if you were actually grabbing something
before passing to the proc like in the other areas it is called which
caused extraneous messages.

Also, mouse drop does not guarantee a mob being passed which was causing
extraneous and confusing messages when dropping wall resin onto a wall
while dragging something.

# Explain why it's good for the game

Confusing and extraneous messages bad.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
fix: Fixed extraneous messages in regards to wall nests
/:cl: