Working Joes can no longer eat (#3828)
# About the pull request

Mechanically prevents eating and being force-fed as a Joe

# Explain why it's good for the game

While not worried about the Joe itself eating the Burger, it physically
does not have anything to feed the burger to (The mouth is mostly
aesthetic with a speaker). As such, it makes no sense that you can feed
food into a Working Joe, and should not be possible.

# Changelog
:cl:
del: Working Joes can no longer be fed.
code: New trait: Cannot eat. Self-explanatory.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>