Emote panel for WJs (#3792)
# About the pull request
Adds an emote panel for WJs.


![image](https://github.com/cmss13-devs/cmss13/assets/41448081/a813a6ec-547a-429a-8d44-c8aefcbb24b5)

In addition, categorizes WJ emotes for ease of viewing in the panel, and
unit tests all emotes to ensure they have a category.

# Explain why it's good for the game
As joe, I have so many cool voicelines! It's a shame I can't remember
the very-uber-specific emote code to say them, which this fixes.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

See section 1

</details>


# Changelog
:cl:
add: Working Joes now have an emote panel to use voice lines, accessible
as an action button.
/:cl:

---------

Co-authored-by: John Doe <johndoe@gmail.com>