Telebaton stun rebalance (#3811)
# About the pull request

This PR lowers the base stun force of a telebaton to 40.

This PR creates multipliers for the stun force based on the user's
policing skill.

The stun force is offset by melee armor rather than energy armor
(regular stun batons). Melee armor is far more abundant on just about
every role and so a little more beef is required for it to be an
effective non-lethal tool.

# Explain why it's good for the game

I'd like to see the roles that have access to a telebaton be able to
effectively handle situations non-lethally as they see fit. A CO or
synth hitting someone a few times with a conveniently storable telebaton
is preferable to a CO insta-BE'ing someone or a synth getting rudely
bullied and relying on CQC luck.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
balance: Telebaton now scales in stunforce effectiveness based on
policing skill
/:cl: