Stops xenos from bypassing dead moving via roller beds (#3810)
# About the pull request

This PR makes it so a roller bed with a dead person on it now is
unmovable. Can still be moved if unconscious or alive so as to mimic
current restrictions.

# Explain why it's good for the game

This has been a longstanding ruling issue that we can now just let be
mechanically prevented. Less rule bloat good.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
fix: Stops xenos from bypassing dead moving via roller beds
/:cl: