Invisible lurker pushing (#3857)
# About the pull request

This PR makes it so when you collide with a human as an invisible lurker
you will become visible.

# Explain why it's good for the game

Being pushed by a fast and barely visible xeno is super cheesy.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
balance: Lurkers now lose their invisibility when they run into a person
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>
Co-authored-by: BeagleGaming1 <56142455+BeagleGaming1@users.noreply.github.com>