Hard delete fixes for predator landmarks and van if their area changes (#3883)
# About the pull request

This PR applies the fixes previously suggested for #3671. That PR was
originally replacing much of the colony in a nightmare insert and it
exposed how predator landmarks do not handle their area being changed
and an issue with the van being replaced. The description made for these
landmarks is based on some area information and that is used to index
them. This PR makes it so the description of the landmark is saved so it
can still be found in the list even if the area changes.

# Explain why it's good for the game

Less hard deletes and more futureproofing.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Drathek, Steelpoint
fix: Fixed possible hardeletes for predator landmarks and vehicles.
Predator teleporation descriptions now do not change if the area is
altered at runtime so they can still be found correctly.
/:cl: