Added preference to disallow de-activation of your current ability (#3904)
# About the pull request

This PR adds a xeno preference where you can disallow the deactivation
of your currently selected ability when re-selecting it again. The
default state of this toggle will still allow ability deactivation since
that is the old behavior.

Locally testing no preferences seem to save so I couldn't confirm that
was working correctly; but I did not introduce a new field to be saved,
just added to the existing one. Maybe they require a DB?

# Explain why it's good for the game

This can help xeno players not accidentally deactivate an ability if
they have no need to deactivate an ability. Generally it should be the
case there is no need to, but maybe someone has multi-binded MMB or
wants to still examine with shift click (if they use shift click
preference).

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/76988376/4324860f-80cd-46ec-9310-dbea6ecbe804)

![image](https://github.com/cmss13-devs/cmss13/assets/76988376/fe8744cb-8a2a-495b-8d96-282417fcacbc)

![image](https://github.com/cmss13-devs/cmss13/assets/76988376/bfde2f29-2e02-4878-97a8-017fedaeb40b)

</details>


# Changelog
:cl: Drathek
qol: Added a preference to disable xeno ability deactivation when
re-selecting the same ability
/:cl: