Adds a CL-exclusive pistol, the ES-4. (#3583)
# About the pull request
This PR adds a CL- and therefore WY - exclusive pistol. The ES-4 is a
Weyland Corp manufactures less-lethal pistol that's essentially a reskin
of the 88 mod 4 that only fires less-lethal projectiles. Sprites by the
fantastic @thwompa
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
The pistol uses distinct blue muzzle flashes for its firing so that it's
easier to distinguish it from normal guns, and it comes straight from
the Aliens RPG. It also looks cool and makes it harder for the CL to get
a lethal firearm straight off the bat.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added the ES-4 electrostatic pistol to the CL's safe.
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>