Trijent Weed Tile Addition (#3890)
# About the pull request

Adds some extra weed tiles near xeno spawns on Desert Dam to prevent
more circumstances of a larva spawning in a trapped area which requires
a xeno to come and rescue the larva.

# Explain why it's good for the game

We don't like trapped larva, at round start. 


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
maptweak: More weed spawns so larva don't spawn in area with no weed to
evolve on.
/:cl:

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>