Gives Sprite To Predator Toolbelt (#3889)
# About the pull request

Creates a unique predator toolbelt with a unique sprite, replacing the
previously edited human toolbelt on the Predator ship. This belt retains
the same equipment and description of the prior one.

# Explain why it's good for the game

Looks slightly more unique and applicable to something a Predator may
carry.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

![Screenshot 2023-07-14 22 08
43](https://github.com/cmss13-devs/cmss13/assets/6595389/f5201793-303d-4866-935e-d2e3d31749c0)

</details>


# Changelog
:cl:
add: Predator toolbelt is now a unique item with its own sprite.
/:cl:

---------

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>
Co-authored-by: Morrow <darthbane97@gmail.com>