Makes the speed for an embryo to burst a config + embryo buff (#3844)
# About the pull request

Changed it from a number to a config.
Changed `larva_gestation_multiplier` to 2, as it only processes once
every 2 seconds, so it should increase the counter by 2 at a time.
Changed an `else if` to `if` as there could never be a situation where
the stage was equal to 4 but not less than 5.

# Explain why it's good for the game

Makes it easier to change, and keeps people guessing what stage they are
at when hugged.

People were operating under the assumption that it was 7.5 minutes.
However, due to `process` only running once every 2 seconds, it ended up
taking twice as long.

Stage 4 was meant to make stasis slightly less effective, but due to a
bug it just never happened.

# Changelog

:cl:
config: Moved time to burst to the config
balance: Fixed time to burst, lowering it from 15 minutes to 7.5 minutes
by default
balance: Stasis bags used on late-stage mobs properly slow growth down
less
/:cl: