Removes Rejuvinate and adds restricted proc logs (#3853)
# About the pull request
Removes R_REJUVINATE as it's only used in two places and we don't need
it.
Adds some alerts where procs are blocked from being force called, and
informs staff if someone attempts to brute force them.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
admin: Removed the R_REJUVINATE permissions flag and replaces instances
where it is used with R_MOD.
admin: Added logs if a restricted proc is proccalled.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>