Adds a new Gin Bottle design, visually different than that of the Vodka. (#3914)
# About the pull request

This PR adds a new Gin Bottle design, as the one currently in game is
visually identical to the Vodka Bottle design, it is inspired in the
Tanqueray London Dry Gin, and reflects the items' description, which
states: "A bottle of high-quality gin, produced in the New London Space
Station." Sprites made by me.



# Explain why it's good for the game

These Changes would benefit the game in a visual manner, as the gin
bottle would be easier to identify from a vodka bottle.


# Testing Photographs and Procedure
  
<details>

https://i.imgur.com/ovpUNab.png

https://i.imgur.com/437rSGS.png

<summary>Screenshots & Videos</summary>

The old Vodka and Gin bottles side by side, and then, the new Gin bottle

</details>


# Changelog

:cl: Kaboomcat
imageadd: Updated the gin bottle sprite
/:cl: