Fix the M56D ignoring collision when dismounting the user (#3929)
# About the pull request

Fixes #3926

It used a force move to dismount from the gun causing these issues. Now,
if you can exit behind you will exit behind the gun.


# Explain why it's good for the game

You could exit into turfs with a density of 1 using this. And you should
not be able to.


# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>

Testing it. With cades and a window frame. To ensure it works.
https://imgur.com/a/eT2LBjW

</details>

# Changelog
:cl:
fix: The M56D no longer let's you violate density code when exiting it.
/:cl: