Donator Kits (#3850)
# About the pull request
Changes all existing donator gear to spawn inside boxes rather than
dumping items into a donator's inventory.
Kits cannot be opened unless by the donator, and can be destroyed with a
right-click by the donator.

Also tidied many of the typepaths that were confusing to make it easier
to understand what the item is.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Makes things tidier.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added donator kit boxes that hold a donator's special gear. They
can be destroyed with a right-click while on the ground, and are locked
to the donator.
code: Tidied up typepaths for many donator items to make it clearer what
or who they are for.
code: Removed the requirement for donators to match their character name
to our config file.
/:cl: