Facehuggers now bypass time of death checks when joining as Larva (#3940)
# About the pull request

Facehuggers ~~(and other ghost roles? (#3939))~~[this functionality will
have to be added in said PR] now bypass the five minute time of death
check, letting them join as a larva via queue or.. join as larva button.

# Explain why it's good for the game

With infection timer at 7.5 minutes, this means that after you die as
facehugger, a large portion of the infection time is spent as
unjoinable, which isnt a problem for a host YOU infected, but for
others, you would get skipped over due to time of death. This allows for
more leniency when playing as low-survivablity ghost roles.

Code added can also be expanded, I am sure the new variable can fit in
other places.

# Changelog
:cl:
add: Facehuggers now bypass time of death checks when being considered
for larva.
code: new variable for observers to handle bypass of time of death
checks
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>