Decreases blood bag IV insertion time (#3950)
# About the pull request

This PR makes it take 1 second instead of 3 seconds (times the surgery
skill multiplier) for blood bag insertion time.

# Explain why it's good for the game

This is taking too long which is severely disincentivizing usage of a
mechanic we want to encourage.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
balance: Decreased blood bag IV insertion time from 3 to 1 second
/:cl: