Fix the first prompt to ghost when nested (#3943)
# About the pull request

This PR fixes an oversight where the initial prompt to ghost when nested
performed ghosting a different way than the ghost verb. Not only does it
now use a TGUI prompt, but the larva queue time will be assigned to the
time they ghosted so they aren't immediately a candidate (the
get_xeno_candidates proc is checking reviveability, but they have no TOD
yet so it skipped that check for nested).

An alternative implementation to this would be to have that proc account
for nested, but I was more inclined to do this this way to update the
prompt to tgui.

# Explain why it's good for the game

More TGUI usage for user prompts, and nested marines should not get
prioritized in the larva queue.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

https://youtu.be/FZTtlhTQPbE

</details>

# Changelog
:cl: Drathek
fix: Fixed the prompt when nested to ghost: Now uses a TGUI prompt and
sets larva queue time.
/:cl: