Fix Xeno Tacmap Persistence (#3957)
# About the pull request

This PR makes it so the xeno tacmap will be closed when the queen dies
on ovi, or leaves ovi. Also fixes the xeno tacmap handling of queen
deletion runtiming.

# Explain why it's good for the game

Fixes #3888 

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Drathek
fix: Fix xeno tacmap staying open when it should be unavailable
/:cl: