Adds a Researcher headset with med/intel (#3956)
# About the pull request

Adds a researcher headset with both medbay and intel channels. Tested,
no issues, one line of code. Good when asked to do so.

# Explain why it's good for the game

Researchers and the Intelligence Officers have a lot of intermixed
gameplay mechanics with limited communication abilities. Both roles do
have access to the Medical channel but it's often filled with medical
issues and can be overlooked. This would increase communication and make
requests to help with intel/corpse retrieval easier on both parties.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Anuv
add: Added a headset to Researchers with the intel channel included
/:cl: