Made Research nightmare insert blast doors no longer be invulnerable (#3788)
# About the pull request

The "fully locked" nightmare insert had blast doors that were:
Unacidable, unpryable and just generally indestructible.
Replaced it with a variant that is non of those, but merely van-proof.

# Explain why it's good for the game

It lead to very strange gameplay where you could destroy the walls, but
not the podlocks. Would be particularly problematic during an LZ2 siege.
I doubt that the intention with this insert was to make Research
completely inaccessible, so I rather made the podlocks destructible
rather than the walls invulnerable too.


https://github.com/cmss13-devs/cmss13/assets/15560820/3d987d21-0fdf-4133-8df4-b65f30036ddf


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
maptweak: Made LV-624's "Fully-locked" Research nightmare insert no
longer have indestructible blast doors.
/:cl:

---------

Co-authored-by: Rigó János <rigojanos@protonmail.com>