Fix bodybags requiring ToD even if unreviveable (#3964)
# About the pull request

This PR adjusts the condition to close a mob in a body bag such that
time of death only matters if the mob is reviveable. Also now doesn't
care if a ghost exists or not since they could have just disconnected
temporarily.

# Explain why it's good for the game

Fixes #3958 

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/76988376/8b80e5db-2c16-4580-86c3-0d442a251b53)

</details>

# Changelog
:cl: Drathek
fix: Fix bodybags not accepting warm (recent) dead bodies even if
unreviveable
/:cl: