Yautja Bracer Charging Rework (#3456)
# About the pull request
Changes how bracers recharge, requiring drain from APC in the field.
Change made as atomisation from #3271 as part of an ongoing project to
make Yautja a more balanced role for what it is.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Balances out power usage and makes things more dynamic. Will tie in with
cloak not using power but being more vulnerable to the methods to force
decloak.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Yautja bracers can now be recharged from APCs.
remove: Yautja bracers no longer naturally recharge on the colony.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>