Yautja Bracer lock toggle & SD message (#3907)
# About the pull request
Allows Yautja to toggle the lock on their bracer, and that of humans or
dead Yautja.
Also adds a message for when SD is cancelled as other Yautja otherwise
don't know if someone blew up or not.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Removes the requirement of cutting off someone's arms if they stole
bracers.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added a toggle for Yautja bracer lock. This also works with the
same mechanic of SDing another Yautja.
add: Added a message to all Yautja if an SD is cancelled.
/:cl: