Yautja Armory and Ship access changes (#3877)
# About the pull request
- Adds locks on the gladiator doors in Yautja Ship, and adds a plasma
rifle armory to the ship locked to Ancient Yautja.
- Reduces amount of healing herbs in the Yautja Ship arena, and makes it
clear they're supposed to stay on the ship. Moving most of them to the
plasma armory.
- Fixes the "consoles" in the Yautja flight deck being the wrong way
around.
- Adds access defines for Yautja ship doors, requiring display of a
bracer ID chip.
- Updates /datum/rank to /datum/yautja_rank to make it clearer what it's
for.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Limits access to powerful gear, but also makes it so equipment usually
admin spawned for abomination hunts is automatically available.

# Testing Photographs and Procedure
I have confirmed my changes work as intended, and particularly
experimented around my change to /obj/proc/allowed() and found no errors
or issues.

# Changelog
:cl:
code: Changes /datum/rank to /datum/yautja_rank to be clearer on what it
is used for.
add: Added access defines for Yautja ship.
add: Added rank identifiers on Yautja equipment presets to help set
access levels on their bracer chips.
maptweak: Changed the doors on the Yautja ship to be their own subtypes,
with certain ones being locked to Yautja or certain Yautja Ranks.
code: Slightly modernised obj/proc/allowed, and also made it check for
access on bracer chips if present.
maptweak: Replaced the unworthy prey section of the Yautja Ship with a
secondary armory for plasma rifles, however the door is inaccessible to
most Yautja. Intended to make it a bit less obvious admins are spawning
them in for the rare times they're used/necessary.
maptweak: Above armory will automatically open if an Abomination is
detected.
maptweak: Reduces amount of herbs on the Yautja Ship.
maptweak: Fixes the "Consoles" orientation in the Yautja flight deck.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>