3rd Party Victory Music (WY, UPP, CLF, USCM) (#3861)
# About the pull request

This adds a "Last Man Standing" condition, where in the event of a Xeno
Minor where all Xenos are dead but humans are still in control of the
ship, we figure out which faction controls the Almayer and play thematic
music for that faction. Currently whichever faction has more than 50% of
the Last Troops Standing, is considered the "winner". Additionally the
Colonial Marines theme song is now officially the USCM hijack comeback
song.

Melancholy1 is the fallback in the case that there is no determined
majority.

This adds:
- Weyland-Yutani
- UPP
- CLF
- USCM (technically)

Note: This only alters what music plays during a Xeno Minor. The round
outcome still counts as a Xeno win both mechanically and statistically.

# Explain why it's good for the game

In the rare event of a botched xeno hijack, it'd be cool to do something
special if a 3rd party manages to secure a USCM Warship.


# Testing Photographs and Procedure
The headcount proc was tested against ERT members making sure that
faction count was accurate.
The music selector was tested by forcing round_status to Xeno_Minor and
verifying that we toggle between whomever is the majority.
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added unique, faction dependent music to Round End in the event
that Xenos lose post hijack.
/:cl:

---------

Co-authored-by: Ben <91219575+Ben10083@users.noreply.github.com>
Co-authored-by: harryob <me@harryob.live>