laser engraves and rewrites fountain pen flavor  + adds detailed lore (#3796)
I've received a lot of positive feedback for this pen so since it is
semi-important now I went ahead and took the time to get out some
extended lore for this item and re-did the original description flavor
text now that I decided it was made by ARMAT.

To-do maybe :
- [ ] Make the pen resistant to acid to comply with the new lore. (not
100% sure yet but would be funny if the only thing left over from a
command member dying is their pen)
- [X] This PR or next but I want to add a laser engraving with the
owner's name directly on the pen.
- [ ] Probably future PR but now that engravings are here I want to make
it so that laser engravings don't happen unless you are command staff


![dreamseeker_w51xOzqhf0](https://github.com/cmss13-devs/cmss13/assets/24533979/275b6b2e-3cc3-47ec-bde7-9988996c16b4)

![dreamseeker_rfZMGoUh14](https://github.com/cmss13-devs/cmss13/assets/24533979/b9732303-ba4a-4535-92ed-3bcc7fe51e33)

Update
----
There is now a laser engraving on the owner of the pen. Here is how that
looks in-game:


![dreamseeker_vUqjoQxujG](https://github.com/cmss13-devs/cmss13/assets/24533979/10659427-b854-44fe-8cbe-10e528f36575)


:cl: Hopek
add: Changed the description flavortext of the fountain pen to match the
new lore. There is now a laser engraving of the owner on the pen itself.
add: Added a detailed lore description for the fountain pen.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>