Gets rid of GBP (#3978)
# About the pull request
Removes the GBP workflow and any mention of it.

# Explain why it's good for the game
To quote Morrow, 
<img width="228" alt="image"
src="https://github.com/cmss13-devs/cmss13/assets/41448081/6f74e888-584c-4b60-b024-1b1e37c0a97d">

CM is a not-massive codebase in how many consistent contributors we
have, so it doesn't feel necessary to have to track the PRs people make
with the numerical system of GBP. If maintainers don't care about it,
it's better-off to remove the workflow to free up an infinitesimal
amount of runner time.

Co-authored-by: John Doe <johndoe@gmail.com>